[FIX] Emojis wont'get rendered at the preview in the thread
